Add functionality to remove dictionaries
The Dictionaries dialog window now has a button that allows the program
to delete dictionaries. Before this, it was necessary to go the
resources folder and to delete dictionaries manually as dictionaries are
simply folders.

*Add button for deleting dictionaries
*Update view to exclude deleted dictionaries

A possible problem is that the program does not verify that the user
wants to delete the dictionaries. As soon as the delete button is
clicked, the folder and all of its contents are deleted without
notification.