Fix save functionality for searched terms
The program would crash when trying to save a definition to a term that
had been accessed through the search box but had not been clicked. Not
setting a searched item as the current item caused the program to crash
because the save function must know which is the current item in order
to save its definition.

*Set searched item as the current item

Fortuitously, setting a searched item as the current item also leads the
user to the location of the term in the term list. This is useful in large
lists that need scrolling.